Fix Windows build of defaults.cc.
BUG=None
R=hpayer@chromium.org

Review URL: https://codereview.chromium.org/24762002

Patch from Ross McIlroy <rmcilroy@chromium.org>.

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@16968 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/defaults.cc
@@ -40,7 +40,7 @@ bool ConfigureResourceConstraintsForCurrentPlatform(
   }
 
   uint64_t physical_memory = i::OS::TotalPhysicalMemory();
-  uintptr_t lump_of_memory = (i::kPointerSize / 4) * i::MB;
+  int lump_of_memory = (i::kPointerSize / 4) * i::MB;
 
   // The young_space_size should be a power of 2 and old_generation_size should
   // be a multiple of Page::kPageSize.
